Add seed Functionality for Resources
The seed functionality is done using Axios to make requests, and can
register a user if needed; otherwise, it'll sign in with the hard coded
credentials for the demo.
An environment variable was added for where the sample data is extracted
from and parsed into the application.
Additionally, simplified the Box and Recipe models.