Fixed Seed Script and Add Info User Routes
The new user information routes exclude the password hash to secure them
from other users. General user details (name, email, and picture) are
necessary for the API to display details for recipes and more.

Alternatives include to populating the details of a user into recipes
when requested, or re-enabling CORS protection.