Update Recipes to Use Custom API
Additionally needed to create a UserContext for simplicity. Probably not
the most efficient; however, I'm more focused on getting it working for
the time being. Like many say, get it working, then make it fast. Do not
try to prematurely optimize. There's an intermediate step; however, that
is the primary gist.